refactor(domain): organize domain by game under games/ and shared/
Engines, bots and rules move from activity/, bot/, maze/ and wordle/ into
games/<game>/ (rps and tic-tac-toe split into bot/ and room/), SeededRng and
stripDiacritics into shared/. Code only one app uses leaves the package:
minesweeper reveal, pong netcode and word-search selection go to the
activity app; room keys, the boggle dictionary, trivia questions, wordle
day guesses and the card game registry go to the api; bicho status goes to
the bot. Battleship's fire payload accepts any number so the engine
reports non-integer coordinates as fire_error instead of the adapter
dropping them.

+++ packages/contracts/src/activity/games/battleship.ts
@@ -82,8 +82,8 @@ export const placeShipsPayloadSchema = z.object({
 });
 
 export const firePayloadSchema = z.object({
-  x: z.number().int(),
-  y: z.number().int(),
+  x: z.number(),
+  y: z.number(),
 });
 
 export const clientMessageSchema = z.discriminatedUnion('type', [
